# Improve performance of Series.to_numpy/to_cupy